DVY-98 Separate each line by word before creating items (#101)
Separate each word on a line before trying to grab the price. This solves the case where sometimes the name is so long on a line that the name and price are scanned together as one thing. It also allows us to put placeholder emojis in-front of prices that come in without names on the line.